requirements.txt
Django==3.1.4
django-filter==2.4.0
djangorestframework==3.12.2
drf-yasg==1.20.0
numpy==1.19.4
pandas==1.1.5
